Refactor common code to base class
results_processor.py is the base class for comp_mngr_results and comp_organizer_results
